refactor(plugin): move types to gatsby-image
