+++ packages/gatsby-transformer-cloudinary/gatsby-image/types.js
@@ -1,0 +1,43 @@
+exports.gatsbyImageTypes = `
+    type CloudinaryAsset implements Node @dontInfer {
+      fixed(
+        base64Width: Int
+        base64Transformations: [String!]
+        chained: [String!]
+        height: Int
+        transformations: [String!]
+        width: Int
+        ignoreDefaultBase64: Boolean
+      ): CloudinaryAssetFixed!
+
+      fluid(
+        base64Width: Int
+        base64Transformations: [String!]
+        chained: [String!]
+        maxWidth: Int
+        transformations: [String!]
+        ignoreDefaultBase64: Boolean
+      ): CloudinaryAssetFluid!
+    }
+
+    type CloudinaryAssetFixed {
+      aspectRatio: Float
+      base64: String
+      height: Float
+      src: String
+      srcSet: String
+      tracedSVG: String
+      width: Float
+    }
+
+    type CloudinaryAssetFluid {
+      aspectRatio: Float!
+      base64: String
+      presentationHeight: Float
+      presentationWidth: Float
+      sizes: String!
+      src: String!
+      srcSet: String!
+      tracedSVG: String
+    }
+  `;
